refactor: P2-T5 - 移除Services中的Observation状态
