+++ LonelyPianistAVP/Models/Immersive/ARTrackingMode.swift
@@ -1,0 +1,5 @@
+enum ARTrackingMode: Equatable {
+    case calibration
+    case practiceBluetoothMIDI
+    case practiceVirtualOrAudio
+}
